Added counting up and down of items in cart. Has error handeling. Checks for how many in stock

+++ Mainpage/Shop/php/deleteCartItem.php
@@ -4,11 +4,13 @@ session_start();
 try{
     
 $itemId=$_POST['itemId'];
+$userId=$_SESSION['userId'];
 
 require("../../../DBConnection/mysql.php");
 
-$stmt = $mysql->prepare("DELETE FROM cart WHERE item_id = :itemId");
+$stmt = $mysql->prepare("DELETE FROM cart WHERE item_id = :itemId AND user_id = :userId");
 $stmt->bindParam(":itemId", $itemId);
+$stmt->bindParam(":userId", $userId);
 $stmt->execute();
 
 $response[] = array(
